requirements.txt
albumentations==1.1.0
Augmentor==0.2.9
Django==3.2.9
dlib==19.22.1
faiss==1.5.3
fire==0.4.0
h5py==3.6.0
imutils==0.5.4
jieba==0.42.1
joblib==1.1.0
lxml==4.6.4
matplotlib==3.5.0
mtcnn==0.1.1
numpy==1.21.2
onnx==1.10.2
onnxruntime==1.9.0
opencv_python_headless==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
pretrainedmodels==0.7.4
pycocotools==2.0.2
pymongo==3.12.1
PyMySQL==1.0.2
PyYAML==6.0
requests==2.25.1
scikit_image==0.18.3
scikit_learn==1.0.1
scipy==1.7.2
skimage==0.0
snownlp==0.12.3
torch==1.10.0
torchsummary==1.5.1
torchvision==0.11.1
tqdm==4.61.2